Move functionality and tests to new controller
