+++ app/plugins/router.plugin.js
@@ -16,6 +16,7 @@ const {
   AuthorisedSystemRoutes,
   BillRunRoutes,
   BillRunInvoiceRoutes,
+  BillRunTransactionRoutes,
   DatabaseRoutes,
   RegimeRoutes,
   RootRoutes,
@@ -30,6 +31,7 @@ const routes = [
   ...AuthorisedSystemRoutes,
   ...BillRunRoutes,
   ...BillRunInvoiceRoutes,
+  ...BillRunTransactionRoutes,
   ...DatabaseRoutes,
   ...TestRoutes,
   ...TransactionRoutes,
