+++ app/routes/bill_run.routes.js
@@ -17,16 +17,6 @@ const routes = [
     handler: PresrocBillRunsController.create
   },
   {
-    method: 'POST',
-    path: '/v1/{regimeId}/billruns/{billRunId}/transactions',
-    handler: NotSupportedController.index
-  },
-  {
-    method: 'POST',
-    path: '/v2/{regimeId}/bill-runs/{billRunId}/transactions',
-    handler: PresrocBillRunsController.createTransaction
-  },
-  {
     method: 'PATCH',
     path: '/v2/{regimeId}/bill-runs/{billRunId}/generate',
     handler: PresrocBillRunsController.generate
@@ -40,11 +30,6 @@ const routes = [
     method: 'GET',
     path: '/v2/{regimeId}/bill-runs/{billRunId}',
     handler: PresrocBillRunsController.view
-  },
-  {
-    method: 'GET',
-    path: '/v2/{regimeId}/bill-runs/{billRunId}/transactions/{transactionId}',
-    handler: PresrocBillRunsController.viewTransaction
   }
 ]
