+++ app/routes/index.js
@@ -4,6 +4,7 @@ const AirbrakeRoutes = require('./airbrake.routes')
 const AuthorisedSystemRoutes = require('./authorised_system.routes')
 const BillRunRoutes = require('./bill_run.routes')
 const BillRunInvoiceRoutes = require('./bill_run_invoice.routes')
+const BillRunTransactionRoutes = require('./bill_run_transaction.routes')
 const DatabaseRoutes = require('./database.routes')
 const RegimeRoutes = require('./regime.routes')
 const RootRoutes = require('./root.routes')
@@ -16,6 +17,7 @@ module.exports = {
   AuthorisedSystemRoutes,
   BillRunRoutes,
   BillRunInvoiceRoutes,
+  BillRunTransactionRoutes,
   DatabaseRoutes,
   RegimeRoutes,
   RootRoutes,
